Update directory structure
Refactor code to remove warnings
