+++ android/brave_java_sources.gni
@@ -51,17 +51,19 @@ brave_java_sources = [
   "../../brave/android/java/org/chromium/chrome/browser/ntp/BraveNewTabPageLayout.java",
   "../../brave/android/java/org/chromium/chrome/browser/ntp/BraveNewTabPageView.java",
   "../../brave/android/java/org/chromium/chrome/browser/ntp/IncognitoNewTabPageView.java",
-  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/BackgroundImage.java",
-  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/NTPImage.java",
-  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/ImageCredit.java",
-  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/SponsoredImageUtil.java",
-  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/NTPUtil.java",
-  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/SponsoredLogo.java",
-  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/SponsoredTab.java",
-  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/NewTabPageListener.java",
-  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/RewardsBottomSheetDialogFragment.java",
+  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/model/NTPImage.java",
+  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/model/BackgroundImage.java",
+  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/model/Wallpaper.java",
+  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/model/TopSite.java",
+  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/model/ImageCredit.java",
+  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/model/SponsoredLogo.java",
+  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/model/SponsoredTab.java",
+  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/util/SponsoredImageUtil.java",
+  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/util/NTPUtil.java",
+  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/util/NewTabPageListener.java",
+  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/util/FetchWallpaperWorkerTask.java",
   "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/NTPBackgroundImagesBridge.java",
-  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/FetchWallpaperWorkerTask.java",
+  "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/RewardsBottomSheetDialogFragment.java",
   "../../brave/android/java/org/chromium/chrome/browser/ntp_background_images/SuperReferralShareDialogFragment.java",
   "../../brave/android/java/org/chromium/chrome/browser/onboarding/BraveRewardsOnboardingFragment.java",
   "../../brave/android/java/org/chromium/chrome/browser/onboarding/BraveRewardsService.java",
