+++ android/java/org/chromium/chrome/browser/ntp_background_images/NTPBackgroundImagesBridge.java
@@ -5,7 +5,6 @@
 
 package org.chromium.chrome.browser.ntp_background_images;
 
-import android.graphics.Bitmap;
 import android.support.annotation.Nullable;
 import java.util.List;
 import java.util.ArrayList;
@@ -14,12 +13,13 @@ import org.chromium.base.ObserverList;
 import org.chromium.base.ThreadUtils;
 import org.chromium.base.annotations.CalledByNative;
 import org.chromium.base.annotations.NativeMethods;
-import org.chromium.chrome.browser.ntp_background_images.NTPImage;
 import org.chromium.chrome.browser.flags.ChromeFeatureList;
 import org.chromium.chrome.browser.BraveFeatureList;
 import org.chromium.chrome.browser.preferences.BravePrefServiceBridge;
 import org.chromium.chrome.browser.profiles.Profile;
-import org.chromium.chrome.browser.ntp_background_images.NewTabPageListener;
+import org.chromium.chrome.browser.ntp_background_images.model.TopSite;
+import org.chromium.chrome.browser.ntp_background_images.model.Wallpaper;
+import org.chromium.chrome.browser.ntp_background_images.util.NewTabPageListener;
 
 public class NTPBackgroundImagesBridge {
     private long mNativeNTPBackgroundImagesBridge;
@@ -27,86 +27,6 @@ public class NTPBackgroundImagesBridge {
             new ObserverList<NTPBackgroundImageServiceObserver>();
     private static List<TopSite> mTopSites = new ArrayList<>();
     private static NewTabPageListener mNewTabPageListener;
-
-    public static class Wallpaper extends NTPImage {
-        private String mImagePath;
-        private int mFocalPointX;
-        private int mFocalPointY;
-        private String mLogoPath;
-        private String mLogoDestinationUrl;
-        private String mThemeName;
-        private boolean mIsSponsored;
-
-        private Wallpaper(String imagePath, int focalPointX, int focalPointY,
-                          String logoPath, String logoDestinationUrl , String themeName, 
-                          boolean isSponsored) {
-            mImagePath = imagePath;
-            mFocalPointX = focalPointX;
-            mFocalPointY = focalPointY;
-            mLogoPath = logoPath;
-            mLogoDestinationUrl = logoDestinationUrl;
-            mThemeName = themeName;
-            mIsSponsored = isSponsored;
-        }
-
-        public String getImagePath() {
-            return mImagePath;
-        }
-
-        public int getFocalPointX() {
-            return mFocalPointX;
-        }
-
-        public int getFocalPointY() {
-            return mFocalPointY;
-        }
-
-        public String getLogoPath() {
-            return mLogoPath;
-        }
-
-        public String getLogoDestinationUrl() {
-            return mLogoDestinationUrl;
-        }
-
-        public String getThemeName() {
-            return mThemeName;
-        }
-
-        public boolean isSponsored() {
-            return mIsSponsored;
-        }
-    }
-
-    public static class TopSite {
-        private String mName;
-        private String mDestinationUrl;
-        private String mBackgroundColor;
-        private String mImagePath;
-
-        private TopSite(String name, String destinationUrl, String backgroundColor, String imagePath) {
-            mName = name;
-            mDestinationUrl = destinationUrl;
-            mBackgroundColor = backgroundColor;
-            mImagePath = imagePath;
-        }
-
-        public String getName() {
-            return mName;
-        }        
-
-        public String getDestinationUrl() {
-            return mDestinationUrl;
-        }
-
-        public String getBackgroundColor() {
-            return mBackgroundColor;
-        }
-
-        public String getImagePath() {
-            return mImagePath;
-        }
-    }
 
     public abstract static class NTPBackgroundImageServiceObserver {
         public abstract void onUpdated();
