+++ android/java/org/chromium/chrome/browser/ntp_background_images/RewardsBottomSheetDialogFragment.java
@@ -29,6 +29,7 @@ import android.support.design.widget.BottomSheetDialog;
 import android.support.design.widget.BottomSheetBehavior;
 import android.view.ViewTreeObserver;
 
+import org.chromium.chrome.R;
 import org.chromium.content_public.browser.LoadUrlParams;
 
 import org.chromium.chrome.browser.BraveAdsNativeHelper;
@@ -39,14 +40,15 @@ import org.chromium.chrome.browser.tab.TabImpl;
 import org.chromium.chrome.browser.util.ConfigurationUtils;
 import org.chromium.chrome.browser.ChromeTabbedActivity;
 import org.chromium.chrome.browser.onboarding.BraveRewardsService;
-import org.chromium.chrome.browser.ntp_background_images.SponsoredImageUtil;
 import org.chromium.chrome.browser.settings.BackgroundImagesPreferences;
 import org.chromium.chrome.browser.customtabs.CustomTabActivity;
 import org.chromium.chrome.browser.BraveRewardsHelper;
 import org.chromium.chrome.browser.tabmodel.TabModel;
 import org.chromium.chrome.browser.tab.TabAttributes;
 import org.chromium.ui.base.DeviceFormFactor;
-import org.chromium.chrome.R;
+import org.chromium.chrome.browser.ntp_background_images.model.SponsoredTab;
+import org.chromium.chrome.browser.ntp_background_images.util.SponsoredImageUtil;
+import org.chromium.chrome.browser.ntp_background_images.util.NewTabPageListener;
 
 import static org.chromium.ui.base.ViewUtils.dpToPx;
 
@@ -64,8 +66,8 @@ public class RewardsBottomSheetDialogFragment extends BottomSheetDialogFragment{
     @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater,
-                             @Nullable ViewGroup container,
-                             @Nullable Bundle savedInstanceState) {
+        @Nullable ViewGroup container,
+        @Nullable Bundle savedInstanceState) {
     	if (getArguments()!=null) {
     		ntpType = getArguments().getInt(SponsoredImageUtil.NTP_TYPE,1);
     	}
