+++ android/java/org/chromium/chrome/browser/ntp_background_images/model/BackgroundImage.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.chromium.chrome.browser.ntp_background_images;
+package org.chromium.chrome.browser.ntp_background_images.model;
 
 public class BackgroundImage extends NTPImage{
     private int imageDrawable;
