+++ android/java/org/chromium/chrome/browser/ntp_background_images/util/NewTabPageListener.java
@@ -3,15 +3,14 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.chromium.chrome.browser.ntp_background_images;
+package org.chromium.chrome.browser.ntp_background_images.util;
 
 import java.util.List;
 
-import org.chromium.chrome.browser.ntp_background_images.NTPBackgroundImagesBridge;
-
+import org.chromium.chrome.browser.ntp_background_images.model.TopSite;
 
 public interface NewTabPageListener {
 	void updateInteractableFlag(boolean isFromBottomSheet);
 	void updateNTPImage();
-	void updateTopSites(List<NTPBackgroundImagesBridge.TopSite> topSites);
+	void updateTopSites(List<TopSite> topSites);
 }
